producer/requirements.txt
kafka-python-ng
faker
pydantic-settings==2.6.1

